#9336: create helper functions for moreh_layerrnorm
