+++ tt_eager/tt_dnn/op_library/moreh_layernorm_backward/gamma_beta_grad/moreh_layernorm_backward_gamma_beta_grad.cpp
@@ -19,50 +19,6 @@
 
 namespace tt {
 
-namespace {
-bool is_hw_dim(uint32_t dim, uint32_t rank) {
-    if (rank == 1 || rank == 2) {
-        return true;
-    }
-    if (rank >= 3) {
-        if (dim >= rank - 2) {
-            return true;
-        }
-    }
-    return false;
-}
-
-uint32_t compute_inner(Shape shape, uint32_t normalized_dims) {
-    uint32_t num_inner = 1;
-    auto rank = shape.rank();
-
-    for (uint32_t i = rank - normalized_dims; i < rank; i++) {
-        auto size = shape[i];
-        if (is_hw_dim(i, rank)) {
-            size = tt::div_up(size, TILE_WIDTH);
-        }
-        num_inner *= size;
-    }
-
-    return num_inner;
-}
-
-uint32_t compute_outer(Shape shape, uint32_t normalized_dims) {
-    uint32_t num_outer = 1;
-    auto rank = shape.rank();
-
-    for (uint32_t i = 0; i < rank - normalized_dims; i++) {
-        auto size = shape[i];
-        if (is_hw_dim(i, rank)) {
-            size = tt::div_up(size, TILE_WIDTH);
-        }
-        num_outer *= size;
-    }
-    return num_outer;
-}
-
-}  // namespace
-
 namespace operations {
 
 namespace primary {
